Enabled import/no-internal-modules in .eslintrc; fixed remaining errors by adding index files where necessary; added eslint-disable rule for error in VoicePlayerToggle, since that component needs a big refactor anyway.
