+++ app/lib/validate/index.js
@@ -1,0 +1,9 @@
+// @flow
+
+import actionArguments from './actionArguments';
+
+const validate = {
+  actionArguments,
+};
+
+export default validate;
