+++ app/modules/contentItems/actions/editPropsForTypeInState.js
@@ -5,7 +5,7 @@ import _ from 'lodash';
 import InvalidArgumentError from 'errors/implementation-errors/InvalidArgumentError';
 import NotYetImplementedError from 'errors/implementation-errors/NotYetImplementedError';
 import UnsupportedOperationError from 'errors/implementation-errors/UnsupportedOperationError';
-import validateActionArguments from 'lib/validate/actionArguments';
+import validate from 'lib/validate';
 
 import * as t from '../actionTypes';
 import { plainTextContentItemTypes, editablePropsForType } from '../model';
@@ -18,7 +18,7 @@ const editPropsForTypeInState = (
   if (!_.includes(plainTextContentItemTypes, contentItem.type)) throw new NotYetImplementedError(`ContentItemType not yet supported`);
   if (!_.isEmpty(_.omit(propsForType, editablePropsForType[contentItem.type]))) throw new InvalidArgumentError(`"props" object contains invalid props for this contentItem type. Type was: "${contentItem.type}". Invalid props were: "${JSON.stringify(_.omit(propsForType, editablePropsForType[contentItem.type]))}"`);
 
-  const validatedPropsForType = validateActionArguments(
+  const validatedPropsForType = validate.actionArguments(
     propsForType,
     editablePropsForType[contentItem.type],
     {
