+++ app/modules/contentItems/components/EditableDisplay/typesToComponentsMap/Heading.js
