+++ app/modules/contentItems/components/EditableDisplay/typesToComponentsMap/index.js
@@ -3,9 +3,10 @@
 import * as React from 'react';
 
 import * as m from '../../../model';
-import Root from '../types/Root';
-import Heading from '../types/Heading';
-import Paragraph from '../types/Paragraph';
+
+import Root from './Root';
+import Heading from './Heading';
+import Paragraph from './Paragraph';
 
 const DummyDisplayComponent = (): React.Node => (
   <p>Not implemented yet.</p>
